+++ src/Dropdowns/Dropdown/Dropdown.tsx
@@ -4,8 +4,8 @@ import { useRef } from 'react'
 import styled, { css, keyframes } from 'styled-components'
 import { compact, isEqual, isNull } from 'lodash'
 import { useMemo } from 'react'
-import { InputText } from '../Inputs/InputText'
-import { Icon, IconType } from '../Icon'
+import { InputText } from '../../Inputs/InputText'
+import { Icon, IconType } from '../../Icon'
 import { DefaultValueTemplate } from '.'
 import TagsValueTemplate from './TagsValueTemplate'
