+++ src/Dropdowns/TagsSelect/TagsSelectValueTemplate.tsx
